monitor/requirements.txt
# Aura-Next Process Monitor — Python dependencies
psutil>=5.9.8
python-socketio[asyncio]>=5.11.2
aiohttp>=3.9.3
